+++ emu/port/devfs-posix.c
@@ -759,7 +759,7 @@ fsdirconv(Chan *c, char *path, char *name, struct stat *s, uchar *va, int nb, in
 		d.gid = u->name;
 	d.muid = "";
 	d.qid = fsqid(s);
-	d.mode = (d.qid.type<<24)|(s->st_mode&0777);
+	d.mode = ((ulong)d.qid.type<<24)|(s->st_mode&0777);	/* LP64: avoid 0x80<<24 sign-extending into bits 32..63 */
 	d.atime = s->st_atime;
 	d.mtime = s->st_mtime;
 	d.length = s->st_size;
